+++ Models/HoaDonBan.cs
@@ -16,7 +16,7 @@ namespace QLCHBanDienThoaiMoi.Models
         public TrangThaiHoaDon TrangThai { get; set; }
         public ICollection<ChiTietHoaDonBan> ChiTietHoaDonBans { get; set; } = new List<ChiTietHoaDonBan>();
         [NotMapped]
-        public decimal ThanhTien => ChiTietHoaDonBans.Sum(ct => ct.GiaBan * (1 - (ct.SanPham?.KhuyenMai?.GiaTri ?? 0) / 100) * ct.SoLuong);
+        public int ThanhTien => (int)ChiTietHoaDonBans.Sum(ct => ct.GiaBan * (1 - (ct.SanPham?.KhuyenMai?.GiaTri ?? 0) / 100) * ct.SoLuong);
     }
     public enum PhuongThucThanhToan
     {
